requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Gemini AI (nuevo SDK)
google-genai>=1.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1

# Async HTTP client
httpx>=0.26.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.14
